Fix to_dict() to handle FVs without a FV Name
If an FV does not contain the optional EFI_FIRMWARE_VOLUME_EXT_HEADER,
then the to_dict() function will crash. The crash occurs when attempting
to access the fvname attribute, because the fvname attribute does not
exist in this case.

The fix is to check if the fvname attribute does not exist. If it does
not, set nameGuid to None.

Signed-off-by: Nate DeSimone <nathaniel.l.desimone@intel.com>